requirements.txt
# SPP Metric Automation Tool Enhanced - Requirements

# Core dependencies
pandas>=1.5.0
snowflake-connector-python>=3.0.0
openpyxl>=3.1.0

# Optional dependencies for enhanced features
xlsxwriter>=3.0.0
pillow>=9.0.0
requests>=2.28.0

# Development and testing
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0
